pyproject.toml
[project]
name = "vector-bt"
version = "0.1.0"
description = "Automated Trading Backtesting Module using VectorBT"
readme = "README.md"
requires-python = ">=3.9,<3.10"
dependencies = [
    "vectorbtpro @ git+ssh://git@github.com/Southdev-tech/vectorbt.pro",
    "pandas>=1.3.0,<2.3.0",
    "numpy>=1.21.0,<1.25.0",
    "matplotlib>=3.4.0,<3.8.0",
    "plotly>=5.0.0,<5.25.0",
    "yfinance>=0.2.0",
    "requests>=2.25.0",
    "python-dateutil>=2.8.0",
    "pytz>=2021.1",
    "dateparser>=1.0.0",
    "exchange-calendars>=4.0.0",
    "ta-lib>=0.4.0",
    "peewee>=3.14.0",
    "schedule>=1.1.0",
    "tqdm>=4.60.0",
    "beautifulsoup4>=4.9.0",
    "curl-cffi>=0.5.0",
    "alpaca-py>=0.40.0",
    "python-dotenv>=1.1.0",
    "numba==0.56.4",
    "llvmlite==0.39.1",
    "psycopg2>=2.9.10",
    "riskfolio-lib>=3.3.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=6.0.0",
    "black>=21.0.0",
    "isort>=5.0.0",
    "flake8>=3.8.0",
]

[tool.uv]
package = false
dev-dependencies = [
    "pytest>=6.0.0",
    "black>=21.0.0",
    "isort>=5.0.0",
    "flake8>=3.8.0",
]

[tool.black]
line-length = 88
target-version = ['py39']

[tool.isort]
profile = "black" 
